+++ blocks/Container.cpp
@@ -48,7 +48,7 @@
             FREQUENT_NEVER : FREQUENT_RARE;
     }
 
-    quint8 Container::Kind() const { return CONTAINER; }
+    int Container::Kind() const { return CONTAINER; }
     int Container::Sub() const { return Block::Sub(); }
     Inventory * Container::HasInventory() { return Inventory::HasInventory(); }
     usage_types Container::Use(Block *) { return USAGE_TYPE_OPEN; }
@@ -88,15 +88,12 @@
         Inventory::SaveAttributes(out);
     }
 
-    Container::Container(const int sub, const quint16 id, const int size) :
+    Container::Container(const int sub, const int id, const int size) :
             Active(sub, id, NONSTANDARD),
             Inventory(size)
     {}
 
-    Container::Container(
-            QDataStream & str,
-            const int     sub,
-            const quint16 id,
+    Container::Container(QDataStream & str, const int sub, const int id,
             const int size)
         :
             Active(str, sub, id, NONSTANDARD),
@@ -128,7 +125,7 @@
         if ( products != nullptr ) {
             for (int i=0; i<products->GetSize(); ++i) {
                 for (int n=0; n<products->GetItem(i)->num; ++n) {
-                    quint16 id = products->GetItem(i)->id;
+                    int id = products->GetItem(i)->id;
                     GetExact(block_manager.NewBlock(
                         block_manager.KindFromId(id),
                         block_manager. SubFromId(id)), i);
@@ -177,7 +174,7 @@
         }
     }
 
-    quint8 Workbench::Kind() const { return WORKBENCH; }
+    int Workbench::Kind() const { return WORKBENCH; }
     int Workbench::Start() const { return 2; }
 
     bool Workbench::Get(Block * const block, const int start) {
@@ -198,9 +195,9 @@
         }
     }
 
-    Workbench::Workbench(const int sub, const quint16 id) :
+    Workbench::Workbench(const int sub, const int id) :
             Container(sub, id, WORKBENCH_SIZE)
     {}
-    Workbench::Workbench(QDataStream & str, const int sub, const quint16 id) :
+    Workbench::Workbench(QDataStream & str, const int sub, const int id) :
             Container(str, sub, id, WORKBENCH_SIZE)
     {}
