+++ blocks/Illuminator.h
@@ -27,14 +27,14 @@ const int MAX_FUEL = SECONDS_IN_DAY;
 class Illuminator : public Active {
     Q_OBJECT
 public:
-    Illuminator(int sub, quint16 id);
-    Illuminator(QDataStream & str, int sub, quint16 id);
+    Illuminator(int sub, int id);
+    Illuminator(QDataStream & str, int sub, int id);
 
-    int     ShouldAct() const override;
-    int     LightRadius() const override;
-    bool    ShouldFall() const override;
-    void    Damage(int dmg, int dmg_kind) override;
-    quint8  Kind() const override;
+    int   ShouldAct() const override;
+    int   LightRadius() const override;
+    bool  ShouldFall() const override;
+    void  Damage(int dmg, int dmg_kind) override;
+    int   Kind() const override;
     Block * DropAfterDamage() override;
     QString FullName() const override;
     usage_types Use(Block *) override;
